+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unreleased
+### Added
+* Added support for extracting textures from `.wimdo` only models to xc3_tex.
+* Added support for extracting textures from `.camdo` models to xc3_tex.
+
 ### Fixed
 * Fixed an issue loading mipmaps for some Xenoblade X textures.